copy_microscopy_datasets_to_userdir.py now uses a cache to store progress and returns immediately.
Long-running plug-in used to outlive the maximum connection length from the client, resulting in the web app announcing an error when in reality the process would finish succesfully. Now the server-side plug-in always returns immediately and runs the actuall process on a separate thread. Subsequent calls from the client now query the status (that is stored in a statically accessed Java hash map). The microscopy-experiment-viewer already uses the new approach. Next commits will address the microscopy-viewer as well.