Change the setShownEdit() for the MICROSCOPY_EXPERIMENT_NAME property to True for new openBIS 16.05 installations.
In version 1 of the microscopy core technlogy, it was set to False. In an openBIS 18.x installation, the master data script will change it programmatically, but this is not possible in openBIS 16.05. Version 2 sets the propery to True so that new installations of openBIS 16.05 will be correct; existing 16.05 installations upgraded to 18.x will be corrected, and only 16.05 installations that get the new microscopy core technology would need to be updated by hand.